+++ infrastructure/outputs.tf
@@ -1,0 +1,3 @@
+output "master_ips" {
+   value = "${join(",", digitalocean_droplet.mesos-master.*.ipv4_address)}"
+}