+++ infrastructure/machines/scripts/bootstrap.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+set -eux
+set -o pipefail
+
+# The sleep 30 in the example above is very important.
+# Because Packer is able to detect and SSH into the instance as soon as SSH is
+# available, Ubuntu actually doesn't get proper amounts of time to initialize.
+# The sleep makes sure that the OS properly initializes.
+sleep 30
+yum update -y