+++ infrastructure/cluster/mesos-master.tf
@@ -24,10 +24,15 @@ resource "digitalocean_droplet" "mesos-master" {
     inline = [
       # start zookeeper
       # TODO version as variable
-      "docker run -d --restart always --name zk --net host -p 2181:2181 -p 2888:2888 -p 3888:3888 jplock/zookeeper:3.4.6",
+      # NOTE test without --net host
+      "docker run -d --restart always --name zk -p 2181:2181 -p 2888:2888 -p 3888:3888 jplock/zookeeper:3.4.6",
 
       # start mesos master
-      "docker run -d --restart always --name master --net host -e MESOS_QUORUM=${var.fleet.masters} -e MESOS_ZK=zk://${digitalocean_droplet.mesos-master.ipv4_address}:2181/mesos -e MESOS_CLUSTER=avengers -e MESOS_HOSTNAME=${digitalocean_droplet.mesos-master.ipv4_address} -p 5050:5050 quay.io/hackliff/mesos.master:cloud mesos-master --ip=${digitalocean_droplet.mesos-master.ipv4_address} --work_dir=/var/lib/mesos/master"
+      "docker run -d --restart always --name master --net host -e MESOS_QUORUM=${var.fleet.masters} -e MESOS_ZK=zk://${digitalocean_droplet.mesos-master.ipv4_address}:2181/mesos -e MESOS_CLUSTER=avengers -e MESOS_HOSTNAME=${digitalocean_droplet.mesos-master.ipv4_address} -p 5050:5050 quay.io/hackliff/mesos.master:cloud mesos-master --ip=${digitalocean_droplet.mesos-master.ipv4_address} --work_dir=/var/lib/mesos/master",
+
+      # start marathon
+      docker run -d --restart always --name marathon -e MARATHON_MASTER=zk://${digitalocean_droplet.mesos-master.ipv4_address}:2181/mesos -e MARATHON_ZK=zk://${digitalocean_droplet.mesos-master.ipv4_address}:2181/marathon -e MARATHON_TASK_LAUNCH_TIMEOUT=300000 -e MARATHON_HOSTNAME=${digitalocean_droplet.mesos-master.ipv4_address} -p 8080:8080 quay.io/hackliff/marathon:cloud
+
     ]
   }
 }