+++ infrastructure/machines/scripts/init-master.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+set -eux
+set -o pipefail
+
+# TODO tag images
+docker pull jplock/zookeeper:${ZK_VERSION}
+docker pull quay.io/hackliff/mesos.master:${MESOS_MASTER_VERSION}
+docker pull quay.io/hackliff/marathon:${MARATHON_VERSION}
+docker pull quay.io/hackliff/chronos:${CHRONOS_VERSION}