+++ changelog.md
@@ -12,6 +12,9 @@ All notable changes to this project will be documented in this file, following [
 
 ### Changed
 - nice ui for bootstrap script
+- re-organize directories
+- improve `build_component` with proper context and tags
+- split scripts into single-purpose ones
 
 ---