+++ changelog.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file, following [
 
 ---
 
+## [0.1.2] - 2015-08-15
+### Added
+- DigitalOcean mesos slave droplet template for packer
+- Makefile
+- basic droplet creation with terraform
+- mesos-slave and master droplets
+
+### Changed
+- nice ui for bootstrap script
+- re-organize directories
+- improve `build_component` with proper context and tags
+- split scripts into single-purpose ones
+- containers provisioning on master
+
+---
+
 ## [0.1.1] - 2015-08-12
 ### Added
 - components Dockerfile